refactor: :construction: Update the project structure
