--- src/app/components/HomeHeader/HomeHeader.module.css
@@ -1,3 +1,0 @@
-.bg {
-  background: url('../../../../public/background.png');
-}
