+++ src/app/services/page.tsx
@@ -1,0 +1,7 @@
+
+
+export default function Services(){
+    return (
+        <h1>Serivices</h1>
+    )
+}
