+++ src/components/Layout/Footer.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-const Footer = () => {
+export const Footer = () => {
   return (
     <footer className="bg-gray-800 text-white py-6">
       <div className="container mx-auto text-center">
@@ -29,4 +29,3 @@ const Footer = () => {
   );
 };
 
-export default Footer;
