+++ src/lib/services.ts
@@ -1,0 +1,36 @@
+export const services = [
+    {
+      title: 'Inteligencia Artificial',
+      description:
+        'Modelos de AI para optimizar procesos y decisiones estratégicas.',
+      image: 'https://via.placeholder.com/100',
+    },
+    {
+      title: 'Desarrollo Web',
+      description:
+        'Creamos sitios web rápidos y modernos, adaptados a tu negocio.',
+      image: 'https://via.placeholder.com/100',
+    },
+    {
+      title: 'Soluciones a Medida',
+      description: 'Herramientas diseñadas para tus necesidades específicas.',
+      image: 'https://via.placeholder.com/100',
+    },
+    {
+      title: 'Despliegue de Infraestructura',
+      description:
+        'Implementamos sistemas escalables y eficientes para tu negocio.',
+      image: 'https://via.placeholder.com/100',
+    },
+    {
+      title: 'Análisis de Datos',
+      description: 'Análisis avanzado para tomar decisiones basadas en datos.',
+      image: 'https://via.placeholder.com/100',
+    },
+    {
+      title: 'Seguridad y Escalabilidad',
+      description:
+        'Soluciones confiables, seguras y preparadas para crecer contigo.',
+      image: 'https://via.placeholder.com/100',
+    },
+  ];
